requirements.txt
# Core web framework
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
starlette==1.0.0

# HTTP clients
httpx[http2]==0.28.1

# Scraping engine (installs Playwright + stealth fingerprinting)
scrapling[all]==0.4.2
patchright==1.58.2
playwright==1.58.0

# Auth / JWT
PyJWT==2.12.1
cryptography==46.0.5

# Env / Config
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1

# Async
anyio==4.13.0

# Parsing helpers
beautifulsoup4==4.14.3
lxml==6.0.2
orjson==3.11.7
reportlab==4.4.1
